requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
WTForms==2.1
flask_marshmallow==0.10.1
marshmallow-sqlalchemy==0.21.0